completed

+++ App Dev Project/backend/controller/todo.controller.js
@@ -1,7 +1,6 @@
 const toDoModel = require("../model/todo.model")
 
 const getToDo = async (req, res) => {
-  //const todo = await toDoModel.find()
   const todo = await toDoModel.find({user_id: req.userId})
   res.json(todo)
 };
@@ -10,13 +9,13 @@ const addToDo = async (req, res) => {
   await toDoModel.create({
     taskName: req.body.taskName,
     taskDone: req.body.taskDone,
-    user_id: req.userId}) //// req.userId ?? why not req.body.user_id
+    user_id: req.userId})
   res.json({message: "Todo added"})
 };
 
 const deleteToDo = async (req, res) => {
-  //await toDoModel.findByIdAndDelete({_id: req.params.id, user_id: req.userId})
-  await toDoModel.findByIdAndDelete({_id: req.params.id})
+  await toDoModel.findByIdAndDelete({_id: req.params.id, user_id: req.userId})
+  // await toDoModel.findByIdAndDelete({_id: req.params.id})
   res.json({message: "Todo deleted"})
 };
